backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
python-dotenv==1.1.1
python-multipart==0.0.20
pydantic==2.9.2
aiohttp==3.12.14
google-cloud-storage==3.2.0
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
requests==2.31.0
sqlite3  # Built-in with Python
